+++ crypto/src/main/java/br/com/pedroxsqueiroz/bt/crypto/services/EntryAmmountGetter.java
@@ -5,9 +5,11 @@ import br.com.pedroxsqueiroz.bt.crypto.utils.config_tools.AnnotadedFieldsConfigu
 import br.com.pedroxsqueiroz.bt.crypto.utils.config_tools.Configurable;
 import lombok.experimental.Delegate;
 
+import java.math.BigDecimal;
+
 public abstract class EntryAmmountGetter extends Configurable {
 
-    public abstract Double get(Wallet wallet);
+    public abstract BigDecimal get(Wallet wallet);
 
     @Delegate(types = Configurable.class)
     private AnnotadedFieldsConfigurer configurer = new AnnotadedFieldsConfigurer(this);
