+++ crypto/src/main/java/br/com/pedroxsqueiroz/bt/crypto/services/exitAmmountGetters/TotalExitAmmountGetter.java
@@ -4,11 +4,13 @@ import br.com.pedroxsqueiroz.bt.crypto.dtos.TradePosition;
 import br.com.pedroxsqueiroz.bt.crypto.services.ExitAmmountGetter;
 import br.com.pedroxsqueiroz.bt.crypto.utils.config_tools.InjectInConfigParam;
 
+import java.math.BigDecimal;
+
 @InjectInConfigParam(alias = "total")
 public class TotalExitAmmountGetter extends ExitAmmountGetter {
 
     @Override
-    public Double get(TradePosition openTrade) {
+    public BigDecimal get(TradePosition openTrade) {
         return openTrade.getEntryAmmount();
     }
 }
